# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Water tracker",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is a documentation of students app"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "https://example.com/api/v1"
    }
  ],
  "tags": [
    {
      "name": "Water",
      "description": "Operations with user watertaking"
    },
    {
      "name": "Auth",
      "description": "Auth operations."
    }
  ],
  "paths": {
    "/user/delete-water/{id}": {
      "delete": {
        "tags": [
          "Water"
        ],
        "summary": "Delete a water intakes",
        "operationId": "delateWater",
        "description": "Delete a water intakes",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "example": "66c4b8bd4719066469ce2d02"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Delete a water intakes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 204
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully deleted water intakes"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    },
    "/user/update-water/{id}": {
      "patch": {
        "tags": [
          "Water"
        ],
        "summary": "Update a water intakes",
        "operationId": "updateWater",
        "description": "Update a water intakes",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "example": "66c4b8bd4719066469ce2d02"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "time": {
                    "description": "Update time in add water",
                    "type": "string",
                    "example": "11:35"
                  },
                  "amount": {
                    "description": "Update amount of users water",
                    "type": "number",
                    "example": 245
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Update a water intakes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully updated a water intakes!"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    },
    "/user/daily-norma": {
      "post": {
        "tags": [
          "Water"
        ],
        "summary": "Set daily norma of user",
        "operationId": "waterId",
        "description": "Set daily norma of water",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "dailyNorm"
                ],
                "properties": {
                  "dailyNorm": {
                    "description": "Daily norma of water",
                    "type": "number",
                    "example": 1500
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Set daily norm of water",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully add a water daily norma!"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    },
    "/user/add-water": {
      "post": {
        "tags": [
          "Water"
        ],
        "summary": "Post daily norma of user",
        "operationId": "dailyNorm",
        "description": "Post daily norma of water",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "dailyNorm"
                ],
                "properties": {
                  "dailyNorm": {
                    "description": "D of water",
                    "type": "number",
                    "example": 1500
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Creates a contact",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully created a contact!"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    },
    "/user/daily-water": {
      "get": {
        "tags": [
          "Water"
        ],
        "summary": "Get list of daily water",
        "operationId": "getDailyWater",
        "description": "Get list of daily water with this endpoint",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully found contacts!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found daily water!"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "type": "object",
                          "items": {
                            "$ref": "#/components/schemas/water"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "responses": {
      "404": {
        "description": "Something went wrong",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "Something went wrong"
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
      "water": {
        "type": "object",
        "required": [
          "_id",
          "userId",
          "amount",
          "time",
          "date"
        ],
        "properties": {
          "_id": {
            "description": "Water id",
            "type": "string",
            "example": "66c4b8bd4719066469ce2d02"
          },
          "userId": {
            "description": "User's id",
            "type": "string",
            "example": "66c4b8bd4719066469ce2d02"
          },
          "amount": {
            "description": "Amount of water",
            "type": "number",
            "example": 1500
          },
          "time": {
            "description": "Time of add water",
            "type": "string",
            "example": "11:35"
          },
          "date": {
            "description": "Date of add water",
            "type": "string",
            "example": "2024-09-01"
          }
        }
      }
    }
  }
}